I rewrote the codes so that everyone could enter their nickname and / or nicknames of their friends so that they will not be kicked

+++ swear_words_checker.py
@@ -13,7 +13,7 @@ for i in swear_words:
     translation = translator.translate(i)
     swear_words_ru.append(translation)
 
-the_creator = 'koroll'
+special_nicknames = [] # you should add your nickname and nicknames of people who shouldn't be kicked
 
 members = {}
 
@@ -32,42 +32,38 @@ class MyClient(commands.Bot):
         print(
             f'{message.author} has just sent a message. Text: {message.content}'
         )
+        if message.author.name not in special_nicknames:
+            for i in swear_words:
+                clear_message = message.content.lower().replace(' ', '')
+                if i in clear_message:
+                    await message.delete()
 
-        for i in swear_words:
-            clear_message = message.content.lower().replace(' ', '')
-            if i in clear_message:
-                await message.delete()
+                    if message.author.name not in members.keys():
+                        members[message.author.name] = 1
+                    elif message.author.name in members.keys():
+                        members[message.author.name] += 1
 
-                if message.author.name not in members.keys():
-                    members[message.author.name] = 1
-                elif message.author.name in members.keys():
-                    members[message.author.name] += 1
-
-                if members[message.author.name] > 5:
-                    await self.kick(message.author)
-                    del members[message.author.name]
+                    if members[message.author.name] > 5:
+                        await self.kick(message.author)
+                        del members[message.author.name]
                     pass
 
-        for i in swear_words_ru:
-            clear_message = message.content.lower().replace(' ', '')
-            if i in clear_message:
-                await message.delete()
-                if message.author not in members:
-                    members[message.author.name] = 1
-                else:
-                    members[message.author.name] += 1
+            for i in swear_words_ru:
+                clear_message = message.content.lower().replace(' ', '')
+                if i in clear_message:
+                    await message.delete()
+                    if message.author not in members:
+                        members[message.author.name] = 1
+                    else:
+                        members[message.author.name] += 1
 
-                if members[message.author.name] > 5:
-                    await self.kick(message.author)
-                    del members[message.author.name]
-                    pass
-
-        if the_creator in members:
-            del members[the_creator]
+                    if members[message.author.name] > 5:
+                        await self.kick(message.author)
+                        del members[message.author.name]
 
     async def on_message_delete(self, message):
         print(
-            f'Welp. {message.author} has just deleted its message. Text: {message.content}'
+            f'{message.author} has just deleted its message. Text: {message.content}'
         )
         print(members)
